--- src/pages/Courses/CourseAttendance/index.jsx
@@ -1,14 +1,0 @@
-import React from "react";
-import CourseAttendanceTop from "./CourseAttendanceTop";
-
-function CourseAttendace() {
-  return (
-    <div>
-      <CourseAttendanceTop />
-      {/* map 함수로 render하기
-      <CourseAttendanceCard /> */}
-    </div>
-  );
-}
-
-export default CourseAttendace;
